requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.1
aiosqlite==0.20.0
fastapi==0.110.0
uvicorn[standard]==0.27.1
